Complete recase renaming of temperature widget files.
