+++ web-ui/src/Components/TemperatureWidget/TemperatureWidget.tsx
